Política de Acceso al Eliminar

+++ app/Http/Controllers/RepositoryController.php
@@ -35,8 +35,12 @@ class RepositoryController extends Controller
         return redirect()->route('repositories.edit', $repository);
     }
 
-    public function destroy(Repository $repository)
+    public function destroy(Request $request, Repository $repository)
     {
+        if ($request->user()->id != $repository->user_id) {
+            abort(403);
+        }
+
         $repository->delete();
 
         return redirect()->route('repositories.index');
